Updates to fix unit tests
