+++ cmd/kops/integration_test.go
@@ -22,18 +22,11 @@ import (
 	"crypto/rsa"
 	"crypto/x509"
 	"encoding/pem"
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/service/ec2"
-	"github.com/aws/aws-sdk-go/service/route53"
-	"github.com/golang/glog"
 	"golang.org/x/crypto/ssh"
 	"io/ioutil"
-	"k8s.io/kops/cloudmock/aws/mockec2"
-	"k8s.io/kops/cloudmock/aws/mockroute53"
 	"k8s.io/kops/cmd/kops/util"
 	"k8s.io/kops/pkg/diff"
-	"k8s.io/kops/upup/pkg/fi/cloudup/awsup"
-	"k8s.io/kops/util/pkg/vfs"
+	"k8s.io/kops/pkg/testutils"
 	"os"
 	"path"
 	"reflect"
@@ -121,7 +114,7 @@ func runTest(t *testing.T, clusterName string, srcDir string, version string, pr
 	factoryOptions := &util.FactoryOptions{}
 	factoryOptions.RegistryPath = "memfs://tests"
 
-	h := NewIntegrationTestHarness(t)
+	h := testutils.NewIntegrationTestHarness(t)
 	defer h.Close()
 
 	h.SetupMockAWS()
@@ -256,7 +249,7 @@ func runTestCloudformation(t *testing.T, clusterName string, srcDir string, vers
 	factoryOptions := &util.FactoryOptions{}
 	factoryOptions.RegistryPath = "memfs://tests"
 
-	h := NewIntegrationTestHarness(t)
+	h := testutils.NewIntegrationTestHarness(t)
 	defer h.Close()
 
 	h.SetupMockAWS()
@@ -336,85 +329,6 @@ func runTestCloudformation(t *testing.T, clusterName string, srcDir string, vers
 			t.Fatalf("cloudformation output differed from expected")
 		}
 	}
-}
-
-type IntegrationTestHarness struct {
-	TempDir string
-	T       *testing.T
-}
-
-func NewIntegrationTestHarness(t *testing.T) *IntegrationTestHarness {
-	h := &IntegrationTestHarness{}
-	tempDir, err := ioutil.TempDir("", "test")
-	if err != nil {
-		t.Fatalf("failed to create temp dir: %v", err)
-	}
-	h.TempDir = tempDir
-
-	vfs.Context.ResetMemfsContext(true)
-
-	return h
-}
-
-func (h *IntegrationTestHarness) Close() {
-	if h.TempDir != "" {
-		if os.Getenv("KEEP_TEMP_DIR") != "" {
-			glog.Infof("NOT removing temp directory, because KEEP_TEMP_DIR is set: %s", h.TempDir)
-		} else {
-			err := os.RemoveAll(h.TempDir)
-			if err != nil {
-				h.T.Fatalf("failed to remove temp dir %q: %v", h.TempDir, err)
-			}
-		}
-	}
-}
-
-func (h *IntegrationTestHarness) SetupMockAWS() {
-	cloud := awsup.InstallMockAWSCloud("us-test-1", "abc")
-	mockEC2 := &mockec2.MockEC2{}
-	cloud.MockEC2 = mockEC2
-	mockRoute53 := &mockroute53.MockRoute53{}
-	cloud.MockRoute53 = mockRoute53
-
-	mockRoute53.MockCreateZone(&route53.HostedZone{
-		Id:   aws.String("/hostedzone/Z1AFAKE1ZON3YO"),
-		Name: aws.String("example.com."),
-		Config: &route53.HostedZoneConfig{
-			PrivateZone: aws.Bool(false),
-		},
-	}, nil)
-	mockRoute53.MockCreateZone(&route53.HostedZone{
-		Id:   aws.String("/hostedzone/Z2AFAKE1ZON3NO"),
-		Name: aws.String("internal.example.com."),
-		Config: &route53.HostedZoneConfig{
-			PrivateZone: aws.Bool(true),
-		},
-	}, []*route53.VPC{{
-		VPCId: aws.String("vpc-234"),
-	}})
-	mockRoute53.MockCreateZone(&route53.HostedZone{
-		Id:   aws.String("/hostedzone/Z3AFAKE1ZOMORE"),
-		Name: aws.String("private.example.com."),
-		Config: &route53.HostedZoneConfig{
-			PrivateZone: aws.Bool(true),
-		},
-	}, []*route53.VPC{{
-		VPCId: aws.String("vpc-123"),
-	}})
-
-	mockEC2.Images = append(mockEC2.Images, &ec2.Image{
-		ImageId:        aws.String("ami-12345678"),
-		Name:           aws.String("k8s-1.4-debian-jessie-amd64-hvm-ebs-2016-10-21"),
-		OwnerId:        aws.String(awsup.WellKnownAccountKopeio),
-		RootDeviceName: aws.String("/dev/xvda"),
-	})
-
-	mockEC2.Images = append(mockEC2.Images, &ec2.Image{
-		ImageId:        aws.String("ami-15000000"),
-		Name:           aws.String("k8s-1.5-debian-jessie-amd64-hvm-ebs-2017-01-09"),
-		OwnerId:        aws.String(awsup.WellKnownAccountKopeio),
-		RootDeviceName: aws.String("/dev/xvda"),
-	})
 }
 
 func MakeSSHKeyPair(publicKeyPath string, privateKeyPath string) error {
