+++ upup/pkg/fi/cloudup/tests/bootstrapchannelbuilder/weave/manifest.yaml
@@ -1,0 +1,63 @@
+kind: Addons
+metadata:
+  creationTimestamp: null
+  name: bootstrap
+spec:
+  addons:
+  - manifest: core.addons.k8s.io/v1.4.0.yaml
+    name: core.addons.k8s.io
+    selector:
+      k8s-addon: core.addons.k8s.io
+    version: 1.4.0
+  - id: pre-k8s-1.6
+    kubernetesVersion: <1.6.0
+    manifest: kube-dns.addons.k8s.io/pre-k8s-1.6.yaml
+    name: kube-dns.addons.k8s.io
+    selector:
+      k8s-addon: kube-dns.addons.k8s.io
+    version: 1.14.4
+  - id: k8s-1.6
+    kubernetesVersion: '>=1.6.0'
+    manifest: kube-dns.addons.k8s.io/k8s-1.6.yaml
+    name: kube-dns.addons.k8s.io
+    selector:
+      k8s-addon: kube-dns.addons.k8s.io
+    version: 1.14.4
+  - manifest: limit-range.addons.k8s.io/v1.5.0.yaml
+    name: limit-range.addons.k8s.io
+    selector:
+      k8s-addon: limit-range.addons.k8s.io
+    version: 1.5.0
+  - id: pre-k8s-1.6
+    kubernetesVersion: <1.6.0
+    manifest: dns-controller.addons.k8s.io/pre-k8s-1.6.yaml
+    name: dns-controller.addons.k8s.io
+    selector:
+      k8s-addon: dns-controller.addons.k8s.io
+    version: 1.7.0
+  - id: k8s-1.6
+    kubernetesVersion: '>=1.6.0'
+    manifest: dns-controller.addons.k8s.io/k8s-1.6.yaml
+    name: dns-controller.addons.k8s.io
+    selector:
+      k8s-addon: dns-controller.addons.k8s.io
+    version: 1.7.0
+  - manifest: storage-aws.addons.k8s.io/v1.6.0.yaml
+    name: storage-aws.addons.k8s.io
+    selector:
+      k8s-addon: storage-aws.addons.k8s.io
+    version: 1.6.0
+  - id: pre-k8s-1.6
+    kubernetesVersion: <1.6.0
+    manifest: networking.weave/pre-k8s-1.6.yaml
+    name: networking.weave
+    selector:
+      role.kubernetes.io/networking: "1"
+    version: 1.9.8-kops.2
+  - id: k8s-1.6
+    kubernetesVersion: '>=1.6.0'
+    manifest: networking.weave/k8s-1.6.yaml
+    name: networking.weave
+    selector:
+      role.kubernetes.io/networking: "1"
+    version: 1.9.8-kops.2
